requirements.txt
requests~=2.26.0
PyYAML~=5.4.1
pytest>=5
requests-mock